Policy Dry Run (#270)
* start work on policies dry runs

* server side for dry runs

* policy server and open api changes

* oh my god this was soo hard

* fix errors

* make buttons better

* package.json fix

* remove comments

+++ packtrain-service/src/main/java/edu/mines/packtrain/models/LateRequest.java
@@ -20,7 +20,7 @@ public class LateRequest {
     private UUID id;
 
     @Column(name = "days_requested")
-    private double daysRequested;
+    private int daysRequested;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "request_type")
